[Refactor] XpLevelCalculator 오브젝트 클래스를 core:util 모듈로 이동
