+++ feature/home/src/main/java/com/ilsangtech/ilsang/feature/home/my/component/UserProfileContent.kt
@@ -38,7 +38,7 @@ import com.ilsangtech.ilsang.designsystem.theme.gray500
 import com.ilsangtech.ilsang.designsystem.theme.primary
 import com.ilsangtech.ilsang.feature.home.BuildConfig
 import com.ilsangtech.ilsang.feature.home.R
-import com.ilsangtech.ilsang.feature.home.my.util.XpLevelCalculator
+import com.ilsangtech.ilsang.core.util.XpLevelCalculator
 
 @Composable
 fun UserProfileContent(
